Implement pchar with right error handling in the right context
