+++ error_handler2.c
@@ -4,7 +4,7 @@ void push_error(stack_t *stack, char **arr, int line_number, FILE *fs);
 void pop_error(stack_t *stack, char **arr, int line_number, FILE *fs);
 void sub_error(stack_t *stack, char **arr, int line_number, FILE *fs);
 void mod_error(stack_t *stack, char **arr, int line_number, FILE *fs);
-
+void pchar_error(stack_t *stack, char **arr, int line_number, FILE *fs);
 /**
  * push_error - handles push error
  * @stack: stack (double linked list)
@@ -69,6 +69,7 @@ void mul_error(stack_t *stack, char **arr, int line_number, FILE *fs)
 	free_list(stack);
 	fclose(fs);
 	exit(EXIT_FAILURE);
+}
 
 /**
  * mod_error - handles stack error on mod
@@ -90,6 +91,26 @@ void mod_error(stack_t *stack, char **arr, int line_number, FILE *fs)
 	else if (stack->n == 0)
 	{
 		fprintf(stderr, "L%d: division by zero\n", line_number);
+		_free(arr);
+		free_list(stack);
+		fclose(fs);
+		exit(EXIT_FAILURE);
+	}
+}
+
+void pchar_error(stack_t *stack, char **arr, int line_number, FILE *fs)
+{
+	if (stack == NULL)
+	{
+		fprintf(stderr, "L%d: can't mod, stack too short\n", line_number);
+		_free(arr);
+		free_list(stack);
+		fclose(fs);
+		exit(EXIT_FAILURE);
+	}
+	else if (stack->n > 127)
+	{
+		fprintf(stderr, "L%d: can't pchar, value out of range\n", line_number);
 		_free(arr);
 		free_list(stack);
 		fclose(fs);
