+++ get_op.c
@@ -10,6 +10,7 @@ func_op get_op(char *opcode);
 func_op get_op(char *opcode)
 {
 	instruction_t handlers[] = {
+		{"pchar", pchar},
 		{"mul", mul},
 		{"div", _div},
 		{"add", add},
