+++ monty.h
@@ -41,6 +41,7 @@ void sub(stack_t **stack, unsigned int line_number);
 void _div(stack_t **stack, unsigned int line_number);
 void mul(stack_t **stack, unsigned int line_number);
 void mod(stack_t **stack, unsigned int line_number);
+void pchar(stack_t **stack, unsigned int line_number);
 
 /**
  * struct instruction_s - opcode and its function
@@ -86,5 +87,6 @@ void zero_error(stack_t *stack, char **arr, int line_number, FILE *fs);
 void mul_error(stack_t *stack, char **arr, int line_number, FILE *fs);
 void mod_error(stack_t *stack, char **arr, int line_number, FILE *fs);
 void push_error(stack_t *stack, char **arr, int line_number, FILE *fs);
+void pchar_error(stack_t *stack, char **arr, int line_number, FILE *fs);
 
 #endif /*_MONTY_H_*/
